Container can provide overriding resources.
Changed the setting so that it can be specified when closing registration.
This allows configuring a "production" container for providing these internal resources, which is not desired.
The other interface options I thought of are harder to convey and remember, so it seems like a reasonable tradeoff.
Perhaps we want to have a different method for this "special" closing so that it's less error_prone.