Resource instances are no longer types.
This was useful when expecting to have a type alias:

class SomeModule:
  a = ModuleResource(int)

And then be able to use SomeModule.a as a type annotation.
But those are not supported by mypy or pycharm's inspections.
Since, we've moved to fake the Resource function signature to Resource(type: t) -> type
but actually return a *Resource instance. So Resources themselves don't play a role in
type checking of user code.