Give up doing complicated type checks for generic and other special types such as unions.
This is clearly a limitation and would allow incompatible provider methods to be accepted by wiring. The problem is just not entirely solvable through the typing annotation system, plus the complexity we'd need to add id equivalent to a whole type checker. Granted, using wiring with incompatible types would not be caught by a type checker and would incur in a runtime error, or worse!